+++ .changeset/modular-linter-architecture.md
@@ -1,0 +1,5 @@
+---
+'@lapidist/design-lint': minor
+---
+
+refactor linter into modular architecture with rule registry, parser service, token tracker, and cache manager
