+++ src/core/token-tracker.ts
@@ -1,0 +1,101 @@
+import type { DesignTokens, LintResult } from './types.js';
+import { extractVarName } from '../utils/token-match.js';
+
+interface UnusedTokenRule {
+  ruleId: string;
+  severity: 'error' | 'warn';
+  ignored: Set<string>;
+}
+
+export class TokenTracker {
+  private allTokenValues: Set<string>;
+  private usedTokenValues = new Set<string>();
+  constructor(private tokens?: DesignTokens) {
+    this.allTokenValues = collectTokenValues(tokens);
+  }
+
+  track(text: string): void {
+    for (const token of this.allTokenValues) {
+      if (this.usedTokenValues.has(token)) continue;
+      if (token.includes('--') || token.startsWith('-')) {
+        if (text.includes(`var(${token})`) || text.includes(token)) {
+          this.usedTokenValues.add(token);
+        }
+      } else if (token.startsWith('#')) {
+        if (text.toLowerCase().includes(token.toLowerCase())) {
+          this.usedTokenValues.add(token);
+        }
+      } else {
+        if (text.includes(token)) this.usedTokenValues.add(token);
+      }
+    }
+  }
+
+  getCompletions(): Record<string, string[]> {
+    const tokens = this.tokens || {};
+    const completions: Record<string, string[]> = {};
+    for (const [group, defs] of Object.entries(tokens)) {
+      if (group === 'deprecations') continue;
+      if (Array.isArray(defs)) {
+        const names = defs.filter((t): t is string => typeof t === 'string');
+        if (names.length) completions[group] = names;
+      } else if (defs && typeof defs === 'object') {
+        const names: string[] = [];
+        for (const val of Object.values(defs)) {
+          const v = typeof val === 'string' ? extractVarName(val) : null;
+          if (v) names.push(v);
+        }
+        if (names.length) completions[group] = names;
+      }
+    }
+    return completions;
+  }
+
+  getUnusedTokenResults(
+    rules: UnusedTokenRule[],
+    configPath = 'designlint.config',
+  ): LintResult[] {
+    const results: LintResult[] = [];
+    for (const { ruleId, severity, ignored } of rules) {
+      const unused = Array.from(this.allTokenValues).filter(
+        (t) => !this.usedTokenValues.has(t) && !ignored.has(t),
+      );
+      if (unused.length) {
+        results.push({
+          filePath: configPath,
+          messages: unused.map((t) => ({
+            ruleId,
+            message: `Token ${t} is defined but never used`,
+            severity,
+            line: 1,
+            column: 1,
+          })),
+        });
+      }
+    }
+    return results;
+  }
+}
+
+function collectTokenValues(tokens?: DesignTokens): Set<string> {
+  const values = new Set<string>();
+  if (!tokens) return values;
+  for (const [group, defs] of Object.entries(tokens)) {
+    if (group === 'deprecations') continue;
+    if (Array.isArray(defs)) {
+      for (const t of defs) {
+        if (typeof t === 'string' && !t.includes('*')) values.add(t);
+      }
+    } else if (defs && typeof defs === 'object') {
+      for (const val of Object.values(defs as Record<string, unknown>)) {
+        if (typeof val === 'string') {
+          const m = val.match(/^var\((--[^)]+)\)$/);
+          values.add(m ? m[1] : val);
+        } else if (typeof val === 'number') {
+          values.add(String(val));
+        }
+      }
+    }
+  }
+  return values;
+}
